fix: Clean up Select file structure and move examples to main react
