+++ modules/react/_examples/stories/mdx/FormsWithFormik.mdx
@@ -1,15 +1,10 @@
 import {ExampleCodeBlock} from '@workday/canvas-kit-docs';
 import {SelectWithFormik} from './examples/SelectWithFormik';
-import {TextInputWithFormik} from './examples/TextInputWithFormik';
 
-<Meta title="Examples/Preview/Forms/With Formik" />
+<Meta title="Examples/Forms/Select With Formik" />
 
 # Canvas Kit Examples
 
 ## Select Using Formik
 
 <ExampleCodeBlock code={SelectWithFormik} />
-
-## Controlled Text Inputs Using Formik
-
-<ExampleCodeBlock code={TextInputWithFormik} />
