+++ modules/react/_examples/stories/mdx/examples/SelectWithFormik.tsx
